Added Auton Types

+++ src/org/usfirst/frc/team747/robot/maps/RobotMap.java
@@ -2,6 +2,6 @@ package org.usfirst.frc.team747.robot.maps;
 
 public class RobotMap {
 	public static double robotLength = 39.25;
-	public static int robotWidth = 36;
+	public static double robotWidth = 36;
 
 }
